feature-NWS-1: Codebase Cleanup
