+++ docker/nginx/default.conf
@@ -7,7 +7,7 @@ server {
     listen 80;
     listen [::]:80;
 
-    server_name api.nebalus.dev;
+    server_name api.nebalus.dev localhost;
 
     root /var/www/public;
     index index.php;
