+++ src/Controller/Linktree/LinktreeApiCreateController.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Nebalus\Webapi\Controller\Referral\Api;
+namespace Nebalus\Webapi\Controller\Linktree;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
-class ReferralApiCreateController
+class LinktreeApiCreateController
 {
     public function __construct()
     {
@@ -13,7 +13,7 @@ class ReferralApiCreateController
 
     public function action(Request $request, Response $response, array $args): Response
     {
-        $response->getBody()->write("Referral Api Create");
+        $response->getBody()->write("Linktree Api Create");
 
         return $response->withStatus(200);
     }
