+++ src/Controller/Referral/ReferralApiDeleteController.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Nebalus\Webapi\Controller\Referral\Api;
+namespace Nebalus\Webapi\Controller\Referral;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
-class ReferralApiUpdateController
+class ReferralApiDeleteController
 {
     public function __construct()
     {
@@ -13,7 +13,7 @@ class ReferralApiUpdateController
 
     public function action(Request $request, Response $response, array $args): Response
     {
-        $response->getBody()->write("Referral Api Update");
+        $response->getBody()->write("Referral Api Delete");
 
         return $response->withStatus(200);
     }
