+++ src/Controller/Referral/ReferralApiUpdateController.php
@@ -1,15 +1,19 @@
 <?php
 
-namespace Nebalus\Webapi\Controller\Linktree\Api;
+namespace Nebalus\Webapi\Controller\Referral;
 
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Http\Message\ServerRequestInterface as Request;
 
-class LinktreeApiReadController
+class ReferralApiUpdateController
 {
+    public function __construct()
+    {
+    }
+
     public function action(Request $request, Response $response, array $args): Response
     {
-        $response->getBody()->write("Linktree Api Read");
+        $response->getBody()->write("Referral Api Update");
 
         return $response->withStatus(200);
     }
