--- src/ValueObject/Referral.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Nebalus\Webapi\ValueObject;
-
-class Referral
-{
-}
